requirements.txt
sentence-transformers==2.2.2
chromadb==0.4.22
torch==2.1.0
Pillow==10.1.0
PyPDF2==3.0.1
transformers==4.36.0

